Protecting api routes

+++ app/api/post/[id]/route.js
@@ -14,14 +14,29 @@ export const GET = async (req, { params }) => {
 }
 
 export const PUT = async (req, { params }) => {
-  const data = await req.json()
-
   try {
+    const session = await getServerSession(authOptions)
+
+    if (!session) {
+      return new Response(
+        JSON.stringify({ message: 'You are not logged in.' }),
+        { status: 400 }
+      )
+    }
+
+    const data = await req.json()
     await connectToDB()
 
     const updatedPost = await Post.findOneAndUpdate({ _id: params.id }, data, {
       new: true,
     })
+
+    if (!updatedPost) {
+      return new Response(
+        { message: 'There is no post with this id' },
+        { status: 400 }
+      )
+    }
 
     return new Response(JSON.stringify(updatedPost), { status: 200 })
   } catch (error) {
@@ -31,6 +46,15 @@ export const PUT = async (req, { params }) => {
 
 export const DELETE = async (req, { params }) => {
   try {
+    const session = await getServerSession(authOptions)
+
+    if (!session) {
+      return new Response(
+        JSON.stringify({ message: 'You are not logged in.' }),
+        { status: 400 }
+      )
+    }
+
     await connectToDB()
 
     await Post.deleteOne({ _id: params.id })
